@ feat: widen action IDs to 32 bits
Item action IDs were capped at 65535, but the attribute store and the
OTBM >= 4 attribute map already carry a full 32-bit value, so the limit
was only in our own accessors and dialogs.

- Item::get/setActionID now take uint32_t; setActionID caps at INT32_MAX
  so the value stays round-trippable through the int32_t attribute store.
- The legacy OTBM_ATTR_ACTION_ID attribute is 16-bit on the wire, so
  serialization clamps to 65535 there; OTBM >= 4 maps are unaffected.
- The properties dialogs accept the wider range, and the "must be between
  100 and 65535" check becomes "must be 0 or at least 100".

@

+++ source/map_drawer.cpp
@@ -2854,7 +2854,7 @@ void MapDrawer::WriteTooltip(Item* item, TileTooltipData& data, bool isHouseTile
 	}
 
 	const uint16_t unique = item->getUniqueID();
-	const uint16_t action = item->getActionID();
+	const uint32_t action = item->getActionID();
 	const std::string& text = item->getText();
 	uint8_t doorId = 0;
 	bool hasDoorProperties = false;
